debounce selection handlers
